LinkedList<String> backend interface
Now the FakeFriend periodically .adds String objects to a
LinkedList<String> defined in the main ChatWindow. The ChatWindow
periodically empties the contents of LinkedList<String> into the
textArea.

It's possible that the LinkedList<String> can be replaced by a
BufferedReader or something, but I don't know how that works yet. In the
meantime, this works.